+++ pkg/sharder/sharder.go
@@ -6,8 +6,8 @@ import (
 
 type Sharder interface {
 	// Returns the keys owned by the current instance.
-	GetMyOwnedKeyRange(context.Context) (uint64, uint64, error)
+	GetMyOwnedKeyRange(context.Context) (uint32, uint32, error)
 
 	// Returns true if the key is owned by the current instance.
-	IsMyOwnedKey(context.Context, uint64) error
+	IsMyOwnedKey(context.Context, uint32) error
 }
